Used a separate index.js for component containers
