+++ src/components/logout/index.js
@@ -1,0 +1,6 @@
+// Copyright (c) 2016 Mattermost, Inc. All Rights Reserved.
+// See License.txt for license information.
+
+import LogoutContainer from './logout_container';
+
+export default LogoutContainer;
