+++ src/components/right_sidebar_menu/index.js
@@ -1,26 +1,6 @@
 // Copyright (c) 2016 Mattermost, Inc. All Rights Reserved.
 // See License.txt for license information.
 
-import {bindActionCreators} from 'redux';
-import {connect} from 'react-redux';
+import RightSidebarMenuContainer from './right_sidebar_menu_container';
 
-import {goToFlaggedPosts, goToRecentMentions} from 'actions/navigation';
-import {logout} from 'actions/users';
-
-import RightSidebarMenu from './right_sidebar_menu';
-
-function mapStateToProps(state, ownProps) {
-    return ownProps;
-}
-
-function mapDispatchToProps(dispatch) {
-    return {
-        actions: bindActionCreators({
-            goToFlaggedPosts,
-            goToRecentMentions,
-            logout
-        }, dispatch)
-    };
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(RightSidebarMenu);
+export default RightSidebarMenuContainer;
